pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "vaultkey-cli"
version = "1.0.0"
description = "Military-grade password manager with Argon2id encryption and zero-knowledge architecture"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Dhruv Kulkarni", email = "kulkarnidhruv02@gmail.com"}
]
maintainers = [
    {name = "Dhruv Kulkarni", email = "kulkarnidhruv02@gmail.com"}
]
keywords = [
    "password", 
    "manager", 
    "security", 
    "encryption", 
    "argon2", 
    "cli", 
    "vault",
    "zero-knowledge",
    "military-grade",
    "breach-detection",
    "cryptography"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Security :: Cryptography",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Operating System :: POSIX",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
]
requires-python = ">=3.8"
dependencies = [
    "cryptography>=41.0.0",
    "click>=8.0.0",
    "tabulate>=0.9.0",
    "pyperclip>=1.8.0",
    "argon2-cffi>=21.0.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "build>=0.8.0",
    "twine>=4.0.0",
    "mypy>=1.0.0",
]
security = [
    "bandit>=1.7.0",
    "safety>=2.0.0",
]

[project.scripts]
vaultkey = "vaultkey.cli:cli"
vk = "vaultkey.cli:cli"

[project.urls]
Homepage = "https://github.com/dhruvkulkarni02/vaultkey"
Repository = "https://github.com/dhruvkulkarni02/vaultkey"
Documentation = "https://github.com/dhruvkulkarni02/vaultkey#readme"
"Bug Reports" = "https://github.com/dhruvkulkarni02/vaultkey/issues"
"Security Policy" = "https://github.com/dhruvkulkarni02/vaultkey/blob/main/SECURITY_ENHANCEMENTS.md"
Changelog = "https://github.com/dhruvkulkarni02/vaultkey/releases"

[tool.setuptools.packages.find]
where = ["."]
include = ["vaultkey*"]
exclude = ["tests*", "build*", "dist*"]

[tool.setuptools.package-data]
vaultkey = ["*.py"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.git
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=vaultkey --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pyperclip.*",
    "tabulate.*",
    "argon2.*",
]
ignore_missing_imports = true
